// src/app/modules/register-form/models/register-data.model.ts
export interface UserFormData {
  age: number;
  weight: number;
  height: number;
  gender: string;
  diseases: string[];
  health_preference: string[];
  diet_preference: string[];
  daily_step_count_target: number;
  daily_calories_target: number;
  active: string;
  weight_target: number;
  name: string;
}